+++ api/src/services/colony/colony.service.ts
@@ -42,7 +42,11 @@ export class ColonyService {
     const oldDeputies = [0,0,0,0,0,0,0,0];
     const newDeputies = [0,0,0,0,0,0,0,0];
     const data = await this.colonyRepository.getAccessInfoByID(colonyId, ownerCode, deputyCode);
-    oldOwner = data.owner[0].member_id;
+    if (data.owner.length > 0) {
+      oldOwner = data.owner[0].member_id;
+    } else {
+      oldOwner = 0;
+    }
     try {
       newOwner = await this.memberRepository.findIdByUsername(givenOwner);
       newOwner = newOwner[0].id;
@@ -50,7 +54,9 @@ export class ColonyService {
       newOwner = 0;
     }
     if (newOwner !== 0) {
-      await this.colonyRepository.removeIdFromAssignment(colonyId, oldOwner, ownerCode);
+      if (oldOwner !== 0) {
+        await this.colonyRepository.removeIdFromAssignment(colonyId, oldOwner, ownerCode);
+      }
       await this.colonyRepository.addIdToAssignment(colonyId, newOwner, ownerCode);
     }
     data.deputies.forEach((deputies, index) => {
